Implementing Programmer Sound
You are now able to play an audio file on your pc through a programmer sound. It will clean up the sounds and instances as they are created, so you can play as many as you want. Can add limits in FMOD. Needs lots of cleaning and sanity checking for bad file and broken handles. Good stuff.